+++ src/paco/models/applications.py
@@ -1205,6 +1205,8 @@ class DeploymentPipelineSourceGitHub(DeploymentPipelineStageAction):
     github_owner = FieldProperty(schemas.IDeploymentPipelineSourceGitHub['github_owner'])
     github_repository = FieldProperty(schemas.IDeploymentPipelineSourceGitHub['github_repository'])
     github_access_token = FieldProperty(schemas.IDeploymentPipelineSourceGitHub['github_access_token'])
+    poll_for_source_changes = FieldProperty(schemas.IDeploymentPipelineSourceGitHub['poll_for_source_changes'])
+
 
 @implementer(schemas.IDeploymentPipelineBuildCodeBuild)
 class DeploymentPipelineBuildCodeBuild(DeploymentPipelineStageAction):